rename: stremaing 패키지를 계층 단위로 분리
