+++ music/src/test/java/com/music/service/MusicStreamingServiceTest.java
@@ -1,15 +1,15 @@
 package com.music.service;
 
 import com.music.adaptor.FileStorage;
-import com.music.streaming.dto.LyricsResponse;
-import com.music.streaming.dto.StreamResponse;
+import com.music.dto.streaming.LyricsResponse;
+import com.music.dto.streaming.StreamResponse;
 import com.music.eneity.Lyrics;
 import com.music.eneity.Music;
 import com.music.eneity.constants.LyricsFormat;
 import com.music.eneity.constants.ReleaseStatus;
 import com.music.repository.LyricsRepository;
 import com.music.repository.MusicRepository;
-import com.music.streaming.service.MusicStreamingService;
+import com.music.service.streaming.MusicStreamingService;
 import java.io.InputStream;
 import java.util.Optional;
 import org.junit.jupiter.api.DisplayName;
